requirements/development.txt
#################################################################
# Requirements specific for the development environment
#################################################################

-r base.txt

coverage==4.5.1
django-eml-email-backend==0.1
flake8==3.5.0
# Fabric 1.10.2 -- Fabric has no Python3 support yet.
Fabric3==1.14.post1
# pycrypto==2.6.1
pytest==6.0.2
#pytest-forked==1.1.3
pytest-cov==2.5.1
pytest-django==4.4.0
pytest-rerunfailures==4.1
pytest-sugar==0.9.1
# Deactivating tldr plugin as this caused issues in Shippable
# pytest-tldr==0.1.5
pytest-xdist==1.29.0
PyVirtualDisplay==2.2
selenium==3.12.0
slackclient==1.2.1
Sphinx==1.7.4
#transifex-client==latest
ElasticMock==1.8.1
django-redis==5.0.0
raven==6.8.0
parameterized==0.8.1
webdriver-manager==3.5.1
chromedriver-binary==94.0.4606.61.0
